Updated target frameworks to .NET 8. Changed type for weights from decimal to double. Updated code style to latest. Moved utility classes to separate project, a shared class library (Galaxon.Maui). Fixed some nullable-related stuff.

+++ GymCalc/Drawables/BarDrawable.cs
@@ -7,16 +7,16 @@ namespace GymCalc.Drawables;
 
 public class BarDrawable : GymObjectDrawable
 {
-    public const double HEIGHT = 20;
+    public const decimal HEIGHT = 20;
 
-    /// <inheritdoc />
-    public override double GetWidth()
+    /// <inheritdoc/>
+    public override decimal GetWidth()
     {
         return CalculateWidth();
     }
 
-    /// <inheritdoc />
-    public override double GetHeight()
+    /// <inheritdoc/>
+    public override decimal GetHeight()
     {
         return HEIGHT;
     }
